backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
websockets==14.1
google-genai==1.5.0
python-dotenv==1.0.1
pydantic-settings==2.7.1
